+++ lib/songle-song-uri-parser.rb
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 require "songle/song_uri"
 require "songle/song_uri/mp3_song_uri"
+require "songle/song_uri/numeric_song_uri"
 require "songle/song_uri/nn_song_uri"
 require "songle/song_uri/nn_scheme_song_uri"
 require "songle/song_uri/nn_short_song_uri"
@@ -12,5 +13,5 @@ require "songle/song_uri/yt_short_song_uri"
 require "songle/song_uri/invalid_song_uri_error"
 
 module ::SongleSongURIParser
-  VERSION = "0.0.3"
+  VERSION = "1.1.0"
 end