+++ lib/songle/song_uri/yt_short_song_uri.rb
@@ -41,7 +41,7 @@ module ::Songle::SongURI
     end
 
     ##
-    # @function
+    # @method
     #
     def type
       return TYPE