+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG.md
 
+## v1.1.0 (2019-01-25)
+- Changed return value of `Songle::SongURI#type` method (e.g. from "yt" to "youtube")
+- Supported to the numeric format (e.g. "https://songle.jp/songs/12345")
+
 ## v1.0.0 (2019-01-09)
 - Initial public release